BACKEND/requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
branca==0.8.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
eventlet==0.40.4
fastapi==0.121.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
folium==0.20.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
passlib==1.7.4
pdf2image==1.17.0
pillow==12.0.0
postgrest==2.25.0
propcache==0.4.1
psycopg==3.2.12
psycopg-binary==3.2.12
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.15.1
PyYAML==6.0.3
realtime==2.25.0
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.3.2
xyzservices==2025.11.0
yarl==1.22.0
